fix warning in QtSvg if qreal == float
(cherry picked from commit 305ce9ea67c11826d91c032ead55edfd5e15462f)

+++ src/svg/qsvggenerator.cpp
@@ -261,7 +261,7 @@ public:
                 constantAlpha &= (stops.at(i).second.alpha() == alpha);
 
             if (!constantAlpha) {
-                const qreal spacing = 0.02;
+                const qreal spacing = qreal(0.02);
                 QGradientStops newStops;
                 QRgb fromColor = PREMUL(stops.at(0).second.rgba());
                 QRgb toColor;
